requirements.txt
h5py
joblib
nibabel
numpy
Pillow
scipy
scikit-learn
torch
torchvision
tqdm
pyyaml
